Moving Zuul models away from the Jenkins ones.
They were starting to get mixed up.
