+++ cibyl/models/ci/system_factory.py
@@ -15,7 +15,8 @@
 """
 from enum import Enum
 
-from cibyl.models.ci.system import JobsSystem, ZuulSystem
+from cibyl.models.ci.system import JobsSystem
+from cibyl.models.ci.zuul.system import ZuulSystem
 from cibyl.utils.dicts import subset
