+++ tests/intr/models/ci/test_system.py
@@ -15,7 +15,8 @@
 """
 from unittest import TestCase
 
-from cibyl.models.ci.system import JobsSystem, System, ZuulSystem
+from cibyl.models.ci.system import JobsSystem, System
+from cibyl.models.ci.zuul.system import ZuulSystem
 from cibyl.orchestrator import Orchestrator
