+++ tests/intr/sources/zuul/test_query.py
@@ -18,9 +18,9 @@ from unittest.mock import Mock
 
 from cibyl.models.ci.build import Build
 from cibyl.models.ci.job import Job
-from cibyl.models.ci.pipeline import Pipeline
-from cibyl.models.ci.project import Project
-from cibyl.models.ci.tenant import Tenant
+from cibyl.models.ci.zuul.pipeline import Pipeline
+from cibyl.models.ci.zuul.project import Project
+from cibyl.models.ci.zuul.tenant import Tenant
 from cibyl.sources.zuul.query import handle_query
